Moving the registry to the base backend, and making defining its public API
