+++ ibis/backends/base_sql/ddl.py
@@ -2,7 +2,7 @@ import re
 
 import ibis.expr.datatypes as dt
 import ibis.expr.schema as sch
-from ibis.backends.base_sql import quote_identifier, type_to_sql_string
+from ibis.backends.base import quote_identifier, type_to_sql_string
 from ibis.backends.base_sqlalchemy.compiler import DDL, DML
 
 fully_qualified_re = re.compile(r"(.*)\.(?:`(.*)`|(.*))")
