+++ ibis/backends/bigquery/compiler.py
@@ -14,8 +14,8 @@ import ibis.expr.datatypes as dt
 import ibis.expr.lineage as lin
 import ibis.expr.operations as ops
 import ibis.expr.types as ir
-from ibis.backends import base_sql
-from ibis.backends.base_sql import fixed_arity, literal, reduction, unary
+from ibis.backends.base import (fixed_arity, literal, reduction, unary,
+                                operation_registry)
 from ibis.backends.base_sql.compiler import (
     BaseExprTranslator,
     BaseSelect,
@@ -346,7 +346,7 @@ STRFTIME_FORMAT_FUNCTIONS = {
 
 
 _operation_registry = {
-    **base_sql.operation_registry,
+    **operation_registry,
 }
 _operation_registry.update(
     {
