+++ ibis/backends/spark/ddl.py
@@ -1,5 +1,5 @@
 from ibis.backends.base_sql import ddl as base_ddl
-from ibis.backends.base_sql import quote_identifier
+from ibis.backends.base import quote_identifier
 
 from .compiler import _type_to_sql_string
